feat: add main menu, transfer camera tracking to `TrackerActivity`
