+++ app/src/main/res/layout/bottom_sheet_layout.xml
@@ -15,7 +15,7 @@
     app:behavior_hideable="false"
     app:behavior_peekHeight="36dp"
     app:layout_behavior="com.google.android.material.bottomsheet.BottomSheetBehavior"
-    tools:showIn="@layout/activity_main">
+    tools:showIn="@layout/activity_tracker">
 
     <ImageView
         android:id="@+id/bottom_sheet_arrow"
